requirements.txt
pyspark
pandas
pyarrow
fastparquet
kafka-python
transformers
torch
sqlalchemy
psycopg2-binary
streamlit
apache-airflow==2.6.3